clean up the password dialog slightly.
2005-04-20  Larry Ewing  <lewing@novell.com>

	* src/FlickrExport.cs: clean up the password dialog slightly.

	* src/FlickrRemote.cs: parse the space as a long since it might
	overflow int now that flickr is all growed up. (patch from Adam
	Israel).

+++ src/FlickrRemote.cs
@@ -14,8 +14,8 @@ public class FlickrRemote {
 	string email;
 	string passwd;
 	string username;
-	int limit;
-	int used;
+	long limit;
+	long used;
 	bool pro;
 
 	public bool ExportTags;
@@ -135,10 +135,10 @@ public class FlickrRemote {
 			foreach (System.Xml.XmlNode child in node.ChildNodes) {
 				switch (child.Name) {
 				case "limit":
-					this.limit = int.Parse (child.ChildNodes [0].Value);
+					this.limit = long.Parse (child.ChildNodes [0].Value);
 					break;
 				case "used":
-					this.limit = int.Parse (child.ChildNodes [0].Value);
+					this.limit = long.Parse (child.ChildNodes [0].Value);
 					break;
 				}
 			}
